+++ ios-app/Utils/EncryptionKeyHandler.swift
@@ -1,0 +1,9 @@
+//
+//  EncryptionKeyHandler.swift
+//  ios-app
+//
+//  Created by Karthik on 05/06/20.
+//  Copyright © 2020 Testpress. All rights reserved.
+//
+
+import Foundation
